requirements.txt
langchain
openai
chromadb
tiktoken
sentence-transformers
unstructured
pypdf
streamlit
python-dotenv
langchain-community
langchain-openai